[리뷰 API] 서비스(예외처리), 예외처리(응답코드, 에러메시지), 컨트롤러

+++ src/main/java/com/sunyoungeom/booktalk/service/ReviewService.java
@@ -1,24 +1,27 @@
 package com.sunyoungeom.booktalk.service;
 
 import com.sunyoungeom.booktalk.domain.Review;
+import com.sunyoungeom.booktalk.exception.review.ReviewException;
+import com.sunyoungeom.booktalk.exception.review.ReviewErrorCode;
 import com.sunyoungeom.booktalk.repository.ReviewRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import jakarta.servlet.http.HttpSession;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
+@RequiredArgsConstructor
 public class ReviewService {
+    
     private final ReviewRepository repository;
-
-    @Autowired
-    public ReviewService(ReviewRepository repository) {
-        this.repository = repository;
-    }
+    private final HttpSession session;
 
     public Review createReview(Review review) {
-        // 중복 리뷰 검증(사용자는 하나의 책에 대해 하나의 리뷰만 가능)
+        // 중복 리뷰 검증
         validateDuplicateReview(review);
+        session.setAttribute("currentUser", "작성자");
+
         // 리뷰 저장
         repository.save(review);
         return review;
@@ -27,7 +30,7 @@ public class ReviewService {
     private void validateDuplicateReview(Review review) {
         repository.findByTitleAndAuthor(review.getTitle(), review.getAuthor())
                 .ifPresent(m -> {
-                    throw new IllegalStateException("이미 작성한 리뷰입니다.");
+                    throw new ReviewException(ReviewErrorCode.REVIEW_ALREADY_EXISTS_ERROR.getMessage());
                 });
     }
 
@@ -55,24 +58,37 @@ public class ReviewService {
         return repository.findByAuthor(author);
     }
 
-    public void updateReview(Long id, String content) {
+    public Review updateReview(Long id, String content) {
         // 리뷰 존재 확인
         Review review = existsById(id);
+        // 작성자 확인
+        checkAuthorMatch(review);
 
         review.setContent(content);
         repository.update(review.getId(), content);
+
+        return review;
+    }
+
+    private void checkAuthorMatch(Review review) {
+        String currentUsername = (String) session.getAttribute("currentUser");
+        if (currentUsername == null || !review.getAuthor().equals(currentUsername)) {
+            throw new ReviewException(ReviewErrorCode.INACTIVE_USER_ERROR.getMessage());
+        }
     }
 
     public void deleteReview(Long id) {
         // 리뷰 존재 확인
         Review review = existsById(id);
+        // 작성자 확인
+        checkAuthorMatch(review);
 
         repository.deleteById(id);
     }
 
     private Review existsById(Long id) {
         Review review = repository.findById(id)
-                .orElseThrow(() -> new IllegalStateException("해당 ID의 리뷰를 찾을 수 없습니다."));
+                .orElseThrow(() -> new ReviewException(ReviewErrorCode.REVIEW_NOT_FOUND_ERROR.getMessage()));
         return review;
     }
 }
